EventHandler.init no longer pre-creates Priority Queues
Removed the code to pre-prepare all of the Priority Queues and Stacks from `EventHandler`. This means that only Priority Queues and Stacks existing will be those which actually contain `Eventable` objects waiting to be processed.
Ultimately, this speeds up iteration by eliminating the empty Priority Queues/Stacks. If it is iterated, it contains an `Eventable` to process.