Eliminated possibility of Deadlocks everywhere
Queues, Stacks, and Listeners are now immune from Deadlocking. A local copy of each of the aforementioned are taken in the operative methods that could potentially cause a deadlocked reference to these collections.

I've also streamlined the housekeeping for `nil` listeners. It's now performed casually each time an Event is being processed against each Listener.